fixing the bug that compiles even if the out folder is there

+++ run.bat
@@ -1,5 +1,7 @@
 @echo off
-reg add HKCU\Console /v VirtualTerminalLevel /t REG_DWORD /d 1 /f > nul
-javac -d out shell/main/Main.java
-if errorlevel 1 goto end
-java -cp out shell.main.Main
+if exist out (
+    java -cp out shell.main.Main
+) else (
+    javac -d out shell/main/*.java
+    java -cp out shell.main.Main
+)
